chore: rename vector_store_base to base
