+++ pilot/vector_store/weaviate_store.py
@@ -8,7 +8,7 @@ from weaviate.exceptions import WeaviateBaseError
 from pilot.configs.config import Config
 from pilot.configs.model_config import KNOWLEDGE_UPLOAD_ROOT_PATH
 from pilot.logs import logger
-from pilot.vector_store.vector_store_base import VectorStoreBase
+from pilot.vector_store.base import VectorStoreBase
 
 CFG = Config()
